float version working
