+++ BB3/BB3.ioc
@@ -824,7 +824,7 @@ RCC.VCOInput1Freq_Value=8000000
 RCC.VCOInput2Freq_Value=8000000
 RCC.VCOInput3Freq_Value=4000000
 RCC.VDD_VALUE=3
-SDMMC1.ClockDiv=0
+SDMMC1.ClockDiv=2
 SDMMC1.IPParameters=ClockDiv
 SH.FMC_A16_CLE.0=FMC_A16,A16_1
 SH.FMC_A16_CLE.ConfNb=1
