+++ BB_esp_fw/spiffs_image/map.htm
@@ -7,7 +7,7 @@
 <div id="mapid" style="height: 800px;"></div>
 
 <label for="sim_spd">Simulated speed</label>
-<input type="number" min="1" max="20" value="20" id="sim_spd" name="sim_spd"/>
+<input type="number" min="1" max="200" value="20" id="sim_spd" name="sim_spd"/>
 
 <style>
     .map_pg_icon
